Fixes the support in bmac/btouch to include extra source files to add to the build
We had a historic accident, the basic idea was:

	btouch api-file.cs -s=api-dependencies.cs extra1 extra2

Where api-file.cs was the API contract to build bindings for, the
api-dependecies.cs were enumarations and structures that were required
to compile the api-file.cs contract and the extra1 and extra2 parameters
were files that were added to the blend (generated code from the api,
and the api-dependencies).

At some point we started sharing the code between the core binding
generator and a general purpose binding generator, and we ended up in
a situation where the above command line did not work, and instead
the meaning became:

	btouch api-file.cs -s=api-dependencies.cs api-file2.cs api-file3.cs

That means that all extra source files listed were treated as API files
(which is what we intend in cases like MonoMac or MonoTouch).   But this broke
the scenario that we wanted.

The workaround was to use --sourceonly, which produced a listing of all
generated files and then let the Makefile manually compile the generated
files, the api-dependencies and any extra files.   This is cumbersome, but
this works fine for larger projects like MonoMac and MonoTouch which had
extra processing to do anyways.

This patch introduces the new -x=FILE to allow users to specify the
extra files that need to be included in the build before the final library
is produced, without having to use --sourceonly.