Make it so users do not accidentally create empty NSError objects as those can crash applications when returned to Objective-C for processing.
Objective-C code expects NSError objects to be properly initialized when returned.
In too many cases users return NSError objects that were created with the default
constructor which really should not take place.

We are making this Obsolete as we should remove this in future versions.   There is
only one mild case where you can use this object and it is when passing to Objective-C
methods that will initialize and poke at the internals of an NSError themselves, but
these are rare.