btouch: always use the public Handle accessor to create CGColor and CGPath.
Nobody seems to know what the -e switch is for anymore, and it must not
be specified to generate classes that can be inherited (since otherwise
the generated methods won't get the logic to call self/super based on
actual type of the object). Using the public Handle accessor allows
user-bindings to use those types too, while at the same time not having
to specify the -e switch to btouch.