MidiServices: Large update to expose properties only where they matter.
Although the C API allows all the properties to be used in every context,
they merely return UnknownProperty for cases where the property does not
apply which is frustrating.

Instead I audited every property to ensure that it is only exposed in the
types where they are supported.

The code had to be copy-pasted as many of the properties are available
in three different classes (EndPoints, Device and Entity).