Support delegate methods without default implementation.
The problem is the following:

There are optional protocol methods which - when implemented - do not
have a valid "default" return value.  Or in other words, if you do
implement the method, then you must do something meaningful in it.

Examples are NSTableViewDelegate's
tableView:dataCellForTableColumn:row: and
tableView:viewForTableColumn:row:.  Both are optional, but if you do
implement them, then you must provide a real implementation and not
return null.  Another example is tableView:heightOfRow: (we return 16
in NSTableView.cs, which may not be correct if the user specified a
very large font in Interface Builder).

This happens in MonoMac when using any .NET event on the view: doing
so sets the view's delegate to an auto-generated implementation of
NSTableViewDelegate.

The solution:

Well, as trivial as it may sound, the solution is simply not to
implement these methods - and since .NET doesn't allow any
unimplemented methods, we need to trick the Objective C runtime into
thinking they are.  Which is actually as simple as overriding
NSObject's respondsToSelector:

The generator now has a new [NoDefaultValue] attribute for this.
Note that calling respondsToSelector: is only mandatory for optional
protocol methods, so you cannot use this attribute on any
abstract methods.

We now generate the following in NSTableView.g.cs:

======
	static IntPtr selGetRowHeight = Selector.GetHandle ("tableView:heightOfRow:");
	static IntPtr selGetDataCell = Selector.GetHandle ("tableView:dataCellForTableColumn:row:");
	static IntPtr selGetViewForItem = Selector.GetHandle ("tableView:viewForTableColumn:row:");
	static IntPtr selRespondsToSelector = Selector.GetHandle ("respondsToSelector:");
	[Export ("respondsToSelector:")]
	bool _RespondsToSelector (IntPtr selHandle)
	{
		if (selHandle.Equals (selGetRowHeight))
			return getRowHeight != null;
		if (selHandle.Equals (selGetDataCell))
			return getDataCell != null;
		if (selHandle.Equals (selGetViewForItem))
			return getViewForItem != null;
		return Messaging.bool_objc_msgSendSuper_intptr (SuperHandle, selRespondsToSelector, selHandle);
	}
====