Prepare CAMediaTimingFunction.cs for inclusion in MonoMac
This class contains a useful overload for getting the control point and
should be included in MonoMac but it appears to have been overlooked.

The namespace is renamed for consistency with MonoMac types and the
class will need to be added to the monomac Make.shared as a separate
commit.