CoreAnimation: start fixing the return value for CAKeyFrameAnimation.FromKeyPath
We should have made this method return a CAKeyFrameAnimation, instead it returns
the base class.   So we now show an Obsolete, pointing developers to the new name
and one day we will be able to change the return type.

We need this two-step process, because changing the return value could potentially
break source code compilation (even if the method has always returned a
CAKeyFrameAnimation), some users might have:

CAPropertyAnimation prop = CAKeyFrameAnimation.FromKeyPath (...);