Performance updates to the AudioToolbox bindings.
Since our marshaling needs are so simple, we do not need to use the
Runtime's more advanced marshaling functionality.

This patch drops PtrToStructure and StructureToPtr for copying
structures and replaced it with pointer operations.

This patch also drops Marshal.SizeOf which is also a runtime API and
uses instead sizeof.

Small tune ups also to clear buffer that were identified using
Instruments as being in the hot path for creating audio buffers.