Fix CGBitmapContext reference counting when creating CGColorSpace.
CGBitmapContextGetColorSpace expects the caller to retain and release the CGColorSpace. It is a Get, not a Create, and so owned should be false. See https://developer.apple.com/documentation/coregraphics/1454058-cgbitmapcontextgetcolorspace

I encountered this while porting a large project to MacOS and testing on OS X El Capitan.
Without this, MonoMac does not retain but on garbage collection does release the object. If a CGImage (for example) is using the color space object, on its later garbage collection, it may crash with an assertion !space->is_singleton when *it* tries to release the object.